env variables for capabilities

+++ cirrocumulus/mongo_db.py
@@ -4,11 +4,13 @@ import os
 
 import pandas._libs.json as ujson
 from bson import ObjectId
-from cirrocumulus.abstract_db import AbstractDB
-from cirrocumulus.util import get_email_domain, get_fs
 from pymongo import MongoClient
 
-from .envir import CIRRO_DB_URI, CIRRO_AUTH_CLIENT_ID, CIRRO_JOB_RESULTS
+from cirrocumulus.abstract_db import AbstractDB
+from cirrocumulus.util import get_email_domain, get_fs
+from .envir import CIRRO_DB_URI, CIRRO_AUTH_CLIENT_ID, CIRRO_JOB_RESULTS, SERVER_CAPABILITY_EDIT_DATASET, \
+    SERVER_CAPABILITY_ADD_DATASET, SERVER_CAPABILITY_DELETE_DATASET, SERVER_CAPABILITY_LINKS, SERVER_CAPABILITY_JOBS, \
+    SERVER_CAPABILITY_FEATURE_SETS, SERVER_CAPABILITY_RENAME_CATEGORIES
 from .invalid_usage import InvalidUsage
 
 
@@ -25,6 +27,7 @@ class MongoDb(AbstractDB):
         self.db = self.client.get_default_database()
 
     def category_names(self, email, dataset_id):
+
         self.get_dataset(email, dataset_id)
         collection = self.db.categories
         results = {}
@@ -41,6 +44,8 @@ class MongoDb(AbstractDB):
         return results
 
     def upsert_category_name(self, email, dataset_id, category, original_value, update):
+        if not self.capabilities()[SERVER_CAPABILITY_RENAME_CATEGORIES]:
+            return
         self.get_dataset(email, dataset_id)
         collection = self.db.categories
         key = str(dataset_id) + '-' + str(category) + '-' + str(original_value)
@@ -113,6 +118,8 @@ class MongoDb(AbstractDB):
         return results
 
     def delete_dataset_view(self, email, view_id):
+        if not self.capabilities()[SERVER_CAPABILITY_LINKS]:
+            return
         collection = self.db.views
         doc = collection.find_one(dict(_id=ObjectId(view_id)))
         self.get_dataset(email, doc['dataset_id'])
@@ -125,6 +132,8 @@ class MongoDb(AbstractDB):
         return format_doc(doc)
 
     def upsert_dataset_view(self, email, dataset_id, view_id, name, value):
+        if not self.capabilities()[SERVER_CAPABILITY_LINKS]:
+            return
         self.get_dataset(email, dataset_id)
         collection = self.db.views
         entity_update = {'last_updated': datetime.datetime.utcnow()}
@@ -142,13 +151,6 @@ class MongoDb(AbstractDB):
             collection.update_one(dict(_id=ObjectId(view_id)), {'$set': entity_update})
             return view_id
 
-    def delete_dataset(self, email, dataset_id):
-        self.get_dataset(email, dataset_id, True)
-        collection = self.db.datasets
-        collection.delete_one(dict(_id=ObjectId(dataset_id)))
-        self.db.filters.delete_many(dict(dataset_id=dataset_id))
-        self.db.categories.delete_many(dict(dataset_id=dataset_id))
-
     def is_importer(self, email):
         # TODO check if user can modify dataset
         user = self.db.users.find_one(dict(email=email))
@@ -156,8 +158,22 @@ class MongoDb(AbstractDB):
             raise False
         return user['importer']
 
+    def delete_dataset(self, email, dataset_id):
+        if not self.capabilities()[SERVER_CAPABILITY_DELETE_DATASET]:
+            return
+
+        self.get_dataset(email, dataset_id, True)
+        collection = self.db.datasets
+        collection.delete_one(dict(_id=ObjectId(dataset_id)))
+        self.db.filters.delete_many(dict(dataset_id=dataset_id))
+        self.db.categories.delete_many(dict(dataset_id=dataset_id))
+
     def upsert_dataset(self, email, dataset_id, dataset_name=None, url=None, readers=None, description=None, title=None,
                        species=None):
+        if dataset_id is None and not self.capabilities()[SERVER_CAPABILITY_ADD_DATASET]:
+            return
+        if dataset_id is not None and not self.capabilities()[SERVER_CAPABILITY_EDIT_DATASET]:
+            return
         collection = self.db.datasets
         update_dict = {}
         if dataset_name is not None:
@@ -201,12 +217,16 @@ class MongoDb(AbstractDB):
         return results
 
     def delete_feature_set(self, email, dataset_id, set_id):
+        if not self.capabilities()[SERVER_CAPABILITY_FEATURE_SETS]:
+            return
         collection = self.db.feature_sets
         doc = collection.find_one(dict(_id=ObjectId(set_id)))
         self.get_dataset(email, doc['dataset_id'])
         collection.delete_one(dict(_id=ObjectId(set_id)))
 
     def upsert_feature_set(self, email, dataset_id, set_id, category, name, features):
+        if not self.capabilities()[SERVER_CAPABILITY_FEATURE_SETS]:
+            return
         self.get_dataset(email, dataset_id)
         collection = self.db.feature_sets
         entity_update = {}
@@ -227,8 +247,9 @@ class MongoDb(AbstractDB):
             return set_id
 
     def create_job(self, email, dataset_id, job_name, job_type, params):
+        if not self.capabilities()[SERVER_CAPABILITY_JOBS]:
+            return
         self.get_dataset(email, dataset_id)
-
         collection = self.db.jobs
         return str(collection.insert_one(
             dict(dataset_id=dataset_id, name=job_name, email=email, type=job_type, params=params,
@@ -262,6 +283,8 @@ class MongoDb(AbstractDB):
                               {'$set': dict(annotations=annotations, last_updated=datetime.datetime.utcnow())})
 
     def update_job(self, email, job_id, status, result):
+        if not self.capabilities()[SERVER_CAPABILITY_JOBS]:
+            return
         collection = self.db.jobs
         doc = collection.find_one(dict(_id=ObjectId(job_id)))
         self.get_dataset(email, doc['dataset_id'])
@@ -283,6 +306,8 @@ class MongoDb(AbstractDB):
         collection.update_one(dict(_id=ObjectId(job_id)), {'$set': dict(status=status, result=result)})
 
     def delete_job(self, email, job_id):
+        if not self.capabilities()[SERVER_CAPABILITY_JOBS]:
+            return
         collection = self.db.jobs
         doc = collection.find_one(dict(_id=ObjectId(job_id)), dict(email=1))
         if doc['email'] == email and not doc.get('readonly', False):
